% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxdev-package.R
\docType{package}
\name{coxdev-package}
\alias{coxdev-package}
\title{R Interface to coxdev C++ routines}
\description{
This package provides R functions to cox model C++ code.
}
\author{
Jonathan Taylor, Trevor Hastie, and Balasubramanian Narasimhan
}
\keyword{package}
