% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxdev.R
\name{make_cox_deviance}
\alias{make_cox_deviance}
\title{Make cox deviance object}
\usage{
make_cox_deviance(
  event,
  start = NA,
  status,
  tie_breaking = c("efron", "breslow"),
  weight
)
}
\arguments{
\item{event}{the event vector}

\item{start}{the start vector, if start/stop}

\item{status}{the status vector}

\item{tie_breaking}{default 'efron'}

\item{weight}{the sample weight}
}
\description{
Make cox deviance object
}
