% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxdev.R
\name{make_stratified_cox_deviance}
\alias{make_stratified_cox_deviance}
\title{Make stratified cox deviance object (C++ implementation)}
\usage{
make_stratified_cox_deviance(
  event,
  start = NA,
  status,
  strata,
  tie_breaking = c("efron", "breslow")
)
}
\arguments{
\item{event}{the event vector of times}

\item{start}{the start vector, if start/stop. Use \code{NA} for just
right censored data}

\item{status}{the status vector indicating event or censoring}

\item{strata}{the strata vector (integer or factor)}

\item{tie_breaking}{default 'efron'}
}
\value{
a list of two functions named \code{coxdev} and \code{information}
each of which takes a linear predictor as argument, along with
weights
}
\description{
Creates a stratified Cox deviance calculator using pure C++ implementation
for all strata processing. This is faster than the R-loop version when
there are many strata.
}
\examples{
set.seed(10101)
nobs <- 100
x <- matrix(rnorm(nobs * 5), nobs, 5)
beta <- c(1, -0.5, 0.3, 0, 0)
fx <- x \%*\% beta
hx <- exp(fx)
ty <- rexp(nobs, hx)
tcens <- rbinom(n = nobs, prob = 0.3, size = 1)
strata <- sample(1:3, nobs, replace = TRUE)
cox_strat <- make_stratified_cox_deviance(
  event = ty,
  status = tcens,
  strata = strata,
  tie_breaking = 'efron'
)
result <- cox_strat$coxdev(linear_predictor = fx)
str(result)
}
