% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxdev.R
\name{make_cox_deviance}
\alias{make_cox_deviance}
\title{Make cox deviance object}
\usage{
make_cox_deviance(
  event,
  start = NA,
  status,
  tie_breaking = c("efron", "breslow"),
  weight = rep(1, length(event))
)
}
\arguments{
\item{event}{the event vector of times}

\item{start}{the start vector, if start/stop. Use \code{NA} for just
right censored data}

\item{status}{the status vector indicating event or censoring}

\item{tie_breaking}{default 'efron'}

\item{weight}{the sample weight the vector of sample weights,
default all ones}
}
\value{
a list of two functions named \code{coxdev} and \code{information}
each of which takes a linear predictor as argument, along with
weights
}
\description{
Make cox deviance object
}
\examples{
set.seed(10101)
nobs <- 100; nvars <- 10
nzc <- nvars \%/\% 3
x <- matrix(rnorm(nobs * nvars), nobs, nvars)
beta <- rnorm(nzc)
fx <- x[, seq(nzc)] \%*\% beta / 3
hx <- exp(fx)
ty <- rexp(nobs,hx)
tcens <- rbinom(n = nobs, prob = 0.3, size = 1)
cox_deviance <- make_cox_deviance(event = ty,
                                  status = tcens,
                                  weight = rep(1.0, length(ty)),
                                  tie_breaking = 'efron')
result  <- cox_deviance$coxdev(linear_predictor = fx)
str(result)
tx  <- t(x)
h <- cox_deviance$information(fx)
I <- tx \%*\% h(x)  ## I should be symmetric
cov  <- solve(I)
}
